+++ src/app/shared/custom-dialog/dialog-title.directive.ts
@@ -1,8 +1,5 @@
-import { Directive, ElementRef, Input } from '@angular/core';
+import { Directive, ElementRef } from '@angular/core';
 import { CustomDialogRef } from './custom-dialog-ref';
-
-/** Counter used to generate unique IDs for dialog elements. */
-let dialogElementUid = 0;
 
 /**
  * Directive for marking an element as a title in a dialog content template.
@@ -14,11 +11,11 @@ let dialogElementUid = 0;
   }
 })
 export class DialogTitleDirective {
-  /** Unique id for the dialog title. If none is supplied, it will be auto-generated. */
-  @Input() id: string = `custom-dialog-title-${dialogElementUid++}`;
+  // Unique id for the dialog title.
+  id: string;
 
   constructor(private elementRef: ElementRef, private dialogRef: CustomDialogRef) {
-    this.dialogRef.titleId = this.id;
+    this.id = this.dialogRef.titleId;
   }
 
 }