+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [6.17.6] - 2024-04-17
+## [6.17.8] - 2024-05-06
 ### Fixed
 - An issue with updating the current value of autocompleter when changing the
   search parameter name (without breaking the loading of saved criteria).
@@ -11,6 +11,17 @@ This project follows [Semantic Versioning](http://semver.org/).
   binding strength from the FHIR server. This allows us to display only existing
   values and fixes the fallback used to load values from the server when we can't
   get them from spec definitions.
+
+## [6.17.7] - 2024-04-29
+### Changed
+- Updated Angular to version 17.
+- Updated Angular Material to version 16.
+### Fixed
+- Styles for autocomplete fields in the cart.
+
+## [6.17.6] - 2024-04-24
+### Changed
+- Update autocomplete-lhc version and disable its progress bar.
 
 ## [6.17.5] - 2024-04-16
 ### Changed